Moved cert related functionality into cert.go
